Support Oracle and MySQL JDBC
