+++ turntables-core/src/main/java/io/github/nblxa/turntables/exception/FeedException.java
@@ -1,0 +1,7 @@
+package io.github.nblxa.turntables.exception;
+
+public class FeedException extends RuntimeException {
+  public FeedException(Throwable cause) {
+    super(cause);
+  }
+}
