+++ turntables-core/src/main/java/io/github/nblxa/turntables/io/ThrowingSupplier.java
@@ -1,0 +1,6 @@
+package io.github.nblxa.turntables.io;
+
+@FunctionalInterface
+public interface ThrowingSupplier<T, E extends Throwable> {
+  T get() throws E;
+}
