+++ turntables-core/src/main/java/io/github/nblxa/turntables/io/rowstore/JdbcRowStore.java
@@ -5,9 +5,12 @@ import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import io.github.nblxa.turntables.Settings;
 import io.github.nblxa.turntables.Tab;
 import io.github.nblxa.turntables.Turntables;
+import io.github.nblxa.turntables.exception.FeedException;
+import io.github.nblxa.turntables.exception.IngestionException;
 import io.github.nblxa.turntables.io.Configuration;
 import io.github.nblxa.turntables.io.Feed;
 import io.github.nblxa.turntables.io.NameSanitizing;
+import io.github.nblxa.turntables.io.ThrowingSupplier;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -16,11 +19,6 @@ import java.util.Objects;
 import java.util.concurrent.locks.ReentrantLock;
 
 public class JdbcRowStore implements RowStore {
-  @FunctionalInterface
-  public interface ThrowingSupplier<T, E extends Throwable> {
-    @NonNull
-    T get() throws E;
-  }
 
   @NonNull
   private final ThrowingSupplier<Connection, Exception> connectionSupplier;
@@ -42,8 +40,8 @@ public class JdbcRowStore implements RowStore {
           .protocolFor(conn.getClass(), Connection.class)
           .feed(name, data)
           .accept(conn);
-    } catch (Exception se) {
-      throw new IllegalStateException(se);
+    } catch (Exception e) {
+      throw new FeedException(e);
     }
   }
 
@@ -58,7 +56,7 @@ public class JdbcRowStore implements RowStore {
     ) {
       return Turntables.from(rs);
     } catch (SQLException se) {
-      throw new IllegalStateException(se);
+      throw new IngestionException(se);
     }
   }
 
@@ -70,7 +68,7 @@ public class JdbcRowStore implements RowStore {
           .cleanUp(name, cleanUpAction)
           .accept(conn);
     } catch (Exception se) {
-      throw new IllegalStateException(se);
+      throw new FeedException(se);
     }
   }
