+++ turntables-mysql8/src/main/resources/META-INF/services/io.github.nblxa.turntables.io.IoProtocolProvider
@@ -1,0 +1,1 @@
+io.github.nblxa.turntables.mysql8.MySqlProtocolProvider
