+++ turntables-oracle18-java8/src/test/java/io/github/nblxa/turntables/test/oracle/OracleRule.java
