--- turntables-test-oracle/src/test/resources/META-INF/services/io.github.nblxa.turntables.io.IoProtocolProvider
@@ -1,1 +1,0 @@
-io.github.nblxa.turntables.oracle.OracleProtocolProvider
